Settings: modify routes at a more appropriate time
Previously routes were being defined after the url was parsed to check for a matching route
